src/Cre_milvus/requirements.txt
# 原有依赖
numpy>=1.21.0
scikit-learn>=1.3.0
hdbscan>=0.8.29
pyyaml>=6.0
pymilvus==2.4.4  # Milvus客户端库
elasticsearch>=7.0.0,<8.0.0
redis>=4.0.0
redisvl>=0.0.5
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
pandas>=1.5.0
markdown>=3.5.0
PyPDF2>=3.0.0
requests>=2.31.0
sentence-transformers>=2.2.0
umap-learn>=0.5.4
plotly>=5.17.0
zhipuai>=1.0.0
modelscope>=1.9.0
# ModelScope相关依赖
adaseq>=0.6.0
python-multipart>=0.0.6
torch>=2.0.0
transformers>=4.21.0
nltk>=3.8
jieba>=0.42.1

pillow>=9.0.0
opencv-python>=4.8.0
ftfy>=6.1.1
regex>=2022.7.9

psutil>=5.9.0
towhee>=1.1.0
langchain>=0.1.0
langchain-community>=0.0.20
pymupdf>=1.23.0

# 可选依赖（用于更好的性能）
# accelerate>=0.21.0  # 用于模型加速
# bitsandbytes>=0.41.0  # 用于模型量化

# 注意事项:
# 1. 确保你的redis是7.0+ 版本，因为我用到了Redis向量库redisvl
# 2. 需要运行Milvus服务器 (可以使用Docker: docker run -p 19530:19530 milvusdb/milvus:latest)
# 3. 或者安装并运行本地Milvus服务器
# 
# 安装命令: pip install -r requirements.txt